payment_service/requirements.txt
asgiref==3.8.1
bson==0.5.10
Django==3.1.12
djangorestframework==3.12.4
djongo==1.3.7
pymongo==3.11.4
python-dateutil==2.9.0.post0
pytz==2025.1
six==1.17.0
sqlparse==0.2.4
typing_extensions==4.12.2
tzdata==2025.2